Rename intdist -> integrated_alerts
Rename both the field names and the file names.

Change the main column names to just "alert_date" and "alert_confidence".
