--- pipelines/integrated_alerts/prefect_flows/integrated_alerts_tasks.py
@@ -1,20 +1,0 @@
-from typing import Optional, Tuple
-
-import xarray as xr
-from prefect import task
-
-from pipelines.integrated_alerts import stages
-
-
-@task
-def setup_compute(
-    datasets: Tuple[xr.DataArray, ...],
-    expected_groups,
-    contextual_name: Optional[str] = None,
-) -> Tuple:
-    return stages.setup_compute(datasets, expected_groups, contextual_name)
-
-
-@task
-def postprocess_result(result: xr.DataArray):
-    return stages.create_result_dataframe(result)
