+++ pipelines/integrated_alerts/stages.py
@@ -1,15 +1,73 @@
-from typing import Callable, Optional, Tuple
+from datetime import date
+from typing import Optional, Tuple
 
+import numpy as np
 import pandas as pd
 import xarray as xr
+from dateutil.relativedelta import relativedelta
 
+from pipelines.globals import (
+    country_10m_zarr_uri,
+    pixel_area_10m_zarr_uri,
+    region_10m_zarr_uri,
+    subregion_10m_zarr_uri,
+)
 from pipelines.prefect_flows.common_stages import (
     create_result_dataframe as common_create_result_dataframe,
+    rollup_by_gadm_and_convert_to_aoi,
 )
 
-LoaderType = Callable[[str, Optional[str]], Tuple[xr.Dataset, ...]]
 ExpectedGroupsType = Tuple
-SaverType = Callable[[pd.DataFrame, str], None]
+
+alerts_confidence = {2: "low", 3: "high"}
+
+
+def load_data(
+    zarr_uri: str,
+    contextual_uri: Optional[str] = None,
+) -> Tuple[xr.DataArray, ...]:
+    """Load in the alert Zarr, the GADM zarrs, and possibly a contextual layer zarr"""
+
+    alerts = _load_zarr(zarr_uri)
+
+    # reindex to alerts to avoid floating point precision issues
+    # when aligning the datasets
+    # https://github.com/pydata/xarray/issues/2217.
+    country = _load_zarr(country_10m_zarr_uri).reindex_like(
+        alerts, method="nearest", tolerance=1e-5
+    )
+    country_aligned = xr.align(alerts, country, join="left")[1].band_data
+    region = _load_zarr(region_10m_zarr_uri).reindex_like(
+        alerts, method="nearest", tolerance=1e-5
+    )
+    region_aligned = xr.align(alerts, region, join="left")[1].band_data
+    subregion = _load_zarr(subregion_10m_zarr_uri).reindex_like(
+        alerts, method="nearest", tolerance=1e-5
+    )
+    subregion_aligned = xr.align(alerts, subregion, join="left")[1].band_data
+    pixel_area = _load_zarr(pixel_area_10m_zarr_uri).reindex_like(
+        alerts, method="nearest", tolerance=1e-5
+    ).astype(np.float64)
+    pixel_area_aligned = xr.align(alerts, pixel_area, join="left")[1].band_data
+
+    if contextual_uri is not None:
+        contextual_layer = _load_zarr(contextual_uri).reindex_like(
+            alerts, method="nearest", tolerance=1e-5
+        )
+        contextual_layer_aligned = xr.align(alerts, contextual_layer, join="left")[
+            1
+        ].band_data
+    else:
+        contextual_layer_aligned = None
+
+    return (
+        alerts,
+        country_aligned,
+        region_aligned,
+        subregion_aligned,
+        pixel_area_aligned,
+        contextual_layer_aligned,
+    )
 
 
 def setup_compute(
@@ -17,22 +75,43 @@ def setup_compute(
     expected_groups: Optional[ExpectedGroupsType],
     contextual_column_name: Optional[str] = None,
 ) -> Tuple:
-    """Setup the arguments for the xrarray reduce on natural lands by area"""
-    base_zarr, country, region, subregion, contextual_layer = datasets
+    """Setup the arguments for the xarray reduce on alerts"""
+    (alerts, country,
+     region, subregion,
+     pixel_area, contextual_layer) = datasets
 
-    mask = base_zarr.band_data
+    base_layer = pixel_area
     groupbys: Tuple[xr.DataArray, ...] = (
         country.rename("country"),
         region.rename("region"),
         subregion.rename("subregion"),
+        alerts.alert_date,
+        alerts.confidence,
     )
     if contextual_layer is not None:
-        groupbys = groupbys + (contextual_layer.rename(contextual_column_name),)
+        groupbys = (
+            groupbys[:3]
+            + (contextual_layer.rename(contextual_column_name),)
+            + groupbys[3:]
+        )
 
-    return (mask, groupbys, expected_groups)
+    return (base_layer, groupbys, expected_groups)
 
 
-def create_result_dataframe(alerts_count: xr.DataArray) -> pd.DataFrame:
-    df = common_create_result_dataframe(alerts_count)
-    del df["natural_lands"]
+def create_result_dataframe(alerts_area: xr.DataArray) -> pd.DataFrame:
+    df = common_create_result_dataframe(alerts_area)
+    df.rename(columns={"value": "area_ha"}, inplace=True)
+    df.rename(columns={"confidence": "alert_confidence"}, inplace=True)
+    df.rename(columns={"alert_date": "alert_date"}, inplace=True)
+    df["alert_date"] = df.sort_values(by="alert_date").alert_date.apply(
+        lambda x: date(2014, 12, 31) + relativedelta(days=x)
+    )
+    df["alert_confidence"] = df.alert_confidence.apply(
+        lambda x: alerts_confidence[x]
+    )
+    df = rollup_by_gadm_and_convert_to_aoi(df, ['alert_date', 'alert_confidence'])
     return df
+
+
+def _load_zarr(zarr_uri):
+    return xr.open_zarr(zarr_uri, storage_options={"requester_pays": True})
